fix(wasm): a definition registers only the words its statement writes out
`lower_definition` handed the runtime `Procedure`'s recorded name, parameter
list and body text. But `Lowering::lower_proc` records the *written* body while
compiling the body from a value it may have materialised instead — a
const-mapped `$body`, or a `[subst -nocommands …]` template. `proc make {} {
set body {return hello} ; proc p {x} $body }` really does produce
`Procedure { body_source: Some("${body}"), body: <compiled from "return
hello"> }`. Registering that word reports the wrong `info body`, and any later
run of the source body — a step trace, or a declined entry — evaluates the
substitution in the *procedure's own frame*, where its operands do not exist.

A latent trap rather than a live wrong answer, and the distinction matters: I
could not make it answer wrongly through any tier, because both materialising
paths need `proc_depth > 0` and no procedure-body site is ever proven (those
units lower under `DispatchEntryAssumption::UnknownWorld`, which empties the
analysis). The two are mutually exclusive by coincidence, not by design, and
P5 proper plans to make procedure bodies proven.

So the guard makes the invariant hold by construction: name, parameters and
body must each be written out literally, and the recorded parameter and body
text must be the word's own. Anything else keeps the generic invocation, where
the runtime's `proc` evaluates the word at the call site as Tcl does. The test
lowers the enclosing body under `PristineRegistryWorld` to remove the
coincidence and fails without the guard; its sibling shows a written-out body
still binds under that same proof.

The legacy analysis tier's `tcl_codegen_proc_register` call has the same
unchecked registration and is unreachable for the same reason; it is out of
this issue's scope and noted in the lane document for its own ticket.

No budget row moves.

+++ rust/tcl-compiler/src/codegen/wasm/native_emit.rs
@@ -1542,7 +1542,7 @@ impl Emitter<'_, '_> {
                 // `settle_return` ignores every completion code but `Return`.
                 if *code == CompletionCode::Return {
                     self.i32(PLAIN_RETURN_LEVEL);
-                    self.i32(i64::from(CompletionCode::Ok.as_int()));
+                    self.i32(CompletionCode::Ok.as_int());
                     self.call(self.imports.return_state);
                 }
                 self.i32(code.as_int());
